requirements.txt
dateutil==1.12.0
board==0.0.0.post0
adafruit-circuitpython-neopixel==3.3.7
RPi.GPIO==0.7.0
